requirements.txt
pandas>=1.5.0
psycopg2-binary>=2.9.0
sqlparse>=0.4.3
python-dotenv>=1.0.0
pytest>=7.0.0
requests>=2.28.0